maintenance: add a shared executorService and toChannelAndToken() in client
the shared executorService can be used by other type of clients and it will be shutdown by close().
this allows all threads used in other type of client to be managed by the client.

toChannelAndToken() takes in a channelBuilder and returns a channel and token which is
authenticated by user and password. this allows other type client to create new stub base on
new channel.