Provide async methods for Watch
*Motiviation*

Currently Watch doesn't provide any async versions of methods to listen on watch responses. It also spawns a thread per watch, which is
not needed at all. The implementation can't be used in an environment that potentially set multiple watchers.

*Solution*

Introduce a `listenAsync` method and rewrite the implementation to avoid using a thread per watch.