enable more of error-prone's checks as errors to enforce them
and fix up their impacts; none of which is a glaring major blocker bug,
but logging any runtime exception which could have occurred in a
scheduled task is probably a good idea, and in WatchTest the forgotten
get() on the CompletableFuture returned by delete makes
testWatchOnDelete more stable.  In TxnTest it's more for consistency.
In ClientConnectionManagerTest it seems to intentionally be async, but
we can nevertheless do a get after that latch.await() on the returned
Future instead of just ignoring it.