fix jetcd-all which was broken under OSGi (#382)
and ensure future non-regression by making the
ClientServiceTest actually "do something" now (before
it dit not really).  This can only work if we have an etcd locally running
("sudo systemctl start etcd") ... so we need to actually start a server,
like the tests in jetcd-core do, for this test to pass on CI, and really
prove that jetcd-all could be loaded and used (!) under OSGi.

For this, we use the new launcher artifact introduced in previous commits.

It's a bit of a PITA to use launcher (based on testcontainers) with Pax Exam.
We solve this by introducing the PaxExamWrapperTest which uses JUnitCore to
programmatically launch ClientServiceChecks (formerly ClientServiceTest).
To make it clear that jetcd-launcher does not work under OSGi, we do not
package it as a bundle anymore, so rm maven-bundle-plugin / MANIFEST.