add startsWith() to ByteSequence
related to https://github.com/etcd-io/jetcd/issues/393; I implemented a
startsWith() for ByteSequence in external code, via getByteString(), and
ran into this issue #393.  I could also just have used getBytes() to
implement a startsWith() for ByteSequence, but that would be less
performant.

The hashCode() method was moved just to that it is right after equals,
which is customary.