Close cancelled watchers
A cancelled watcher is effectively dead and will not receive any more
events so it's helpful for users if the watcher gets closed
automatically.

Note that we don't want to reschedule the watcher since users may want
to take corrective action first, such as finding the latest revision if
the watcher was cancelled due to CompactedException.